开发版本，修复bug, 复制模型条件限制，加载文件时把当前用户的所有模型加载

+++ app/service/copy_class.py
@@ -1,4 +1,6 @@
 # -- coding: utf-8 --
+import logging
+
 from app.service.get_model_code import GetModelCode
 from library.file_operation import FileOperation
 from config.omc import omc
@@ -7,33 +9,49 @@ from app.service.load_model_file import LoadModelFile
 
 def CopyClass (copied_class_name, class_name, parent_name):
     exist_class = omc.existClass(parent_name + "." + class_name)
-    if exist_class:
-        return False
     if not parent_name:
-        parent_name = "root"
+        parent_name = "TopLevel"
+    else:
+        parent_information = omc.getClassInformation(parent_name)
+        logging.debug(parent_name)
+        logging.debug(parent_information)
+        if parent_information and parent_information[0] != "package":
+            return False, "Parent node is not a package or root"
+    if exist_class:
+        return False, "Model already exists"
     copy_result = omc.copyClass(copied_class_name, class_name, parent_name)
-    return copy_result
+    return copy_result, "Copy model successfully"
 
 
 def DeleteClass (class_name):
-    result = omc.deleteClass(class_name)
-    return result
-
-
-def SaveClass (class_name, copied_class_name=None, parent_name=None, package_name=None, model_file_path=None,
-               new_model_file_path=None, copy_or_delete="copy"):
-    LoadModelFile(package_name, model_file_path)
-    if copy_or_delete == "copy":
-        result = CopyClass(copied_class_name, class_name, parent_name)
-    elif copy_or_delete == "delete":
-        result = DeleteClass(class_name)
+    exist = omc.existClass(class_name)
+    if exist:
+        result = omc.deleteClass(class_name)
+        if result:
+            return True, "Delete model successfully"
+        else:
+            return False, "Delete model failed"
     else:
-        return False
+        return True, "Delete model successfully"
+
+
+def SaveClass (class_name, copied_class_name=None, parent_name=None, package_name=None,
+               new_model_file_path=None, copy_or_delete="copy", file_name=None):
+    # LoadModelFile(package_name, model_file_path)
+    if copy_or_delete == "copy":
+        result, msg = CopyClass(copied_class_name, class_name, parent_name)
+    elif copy_or_delete == "delete":
+        result, msg = DeleteClass(class_name)
+    else:
+        result = False
+        msg = "Unknown operation"
     if result:
-        data = GetModelCode(package_name)
-        file_name = model_file_path.split("/")[-1]
+        if parent_name:
+            data = GetModelCode(package_name)
+        else:
+            data = GetModelCode(class_name)
+        if not file_name:
+            file_name = new_model_file_path.split("/")[-1]
         path = "/".join(new_model_file_path.split("/")[:-1])
         FileOperation.write_file(path, file_name, data)
-    else:
-        return False
-    return True
+    return result, msg
